+++ crates/hiveweb/src/services/workflow.rs
@@ -143,7 +143,7 @@ pub async fn create(pool: &MySqlPool, meta: CreateMeta) -> Result<Workflow, AppE
     .bind(&meta.identifier)
     .bind(&meta.name)
     .bind(&meta.description)
-    .bind(meta.timeout_ms.unwrap_or(30000))
+    .bind(meta.timeout_ms.unwrap_or(33000))
     .bind(meta.category_id)
     .bind(meta.required_capabilities.as_ref().map(|c| serde_json::to_value(c).unwrap_or(Value::Array(vec![]))))
     .execute(pool)